backend/requirements.txt
fastapi
uvicorn
sqlmodel
alembic
pyjwt
passlib[bcrypt]
python-multipart
pydantic-settings
aiosqlite
asyncpg
